requirements.txt
alembic==1.14.1
email-validator==2.2.0
fastapi==0.115.8
psycopg[binary]==3.2.13
pydantic-settings==2.7.1
PyJWT==2.10.1
python-multipart==0.0.20
SQLAlchemy==2.0.49
uvicorn[standard]==0.34.0
